pyproject.toml
[project]
name = "srt-translator"
version = "0.1.0"
description = "SRT subtitle translation tool"
readme = "README.md"
requires-python = ">=3.13"

# Dependencies from requirements.txt
dependencies = [
    "gradio",
    "llama-index",
    "python-dotenv",
    "openai"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv]
pip-args = [
    "--no-cache-dir",
    "--no-deps"
]